requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12,<3.13
requests>=2.26,<2.27
